Setup base pickup
Pickup script now meant for Kinetic Rigidbody objects.

Currently, movement in hand is default, meaning it passes without collisions. A whole new collision detection and smooth movement system will need to be manually implemented, as it is impossible to apply a force to a kinetic object. (ref. our physics class)

The whole system will have to be implemented in ObjectGrabbable:FixedUpdate()